# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Visiting Card Information Extractor API
 * OpenAPI spec version: 0.2.0
 */
export interface HealthStatus {
  status: string;
}

export interface SignupRequest {
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export type UserProfileRole =
  (typeof UserProfileRole)[keyof typeof UserProfileRole];

export const UserProfileRole = {
  user: "user",
  admin: "admin",
} as const;

export interface UserProfile {
  id: string;
  name: string;
  email: string;
  role: UserProfileRole;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: UserProfile;
}

export interface CardData {
  name: string;
  phones: string[];
  emails: string[];
  company: string;
  designation: string;
  address: string;
  website: string;
}

export interface ExtractCardRequest {
  frontImage: string;
  backImage?: string;
}

export interface ExtractCardResponse {
  id: string;
  data: CardData;
  createdAt: string;
}

export interface CardRecord {
  id: string;
  userId?: string;
  userName?: string;
  data: CardData;
  frontImageUrl?: string;
  backImageUrl?: string;
  createdAt: string;
}

export interface AdminCardsResponse {
  cards: CardRecord[];
  total: number;
  page: number;
  totalPages: number;
}

export interface AdminUserRecord {
  id: string;
  name: string;
  email: string;
  role: string;
  cardCount: number;
  createdAt: string;
}

export interface AdminUsersResponse {
  users: AdminUserRecord[];
  total: number;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export interface SuccessResponse {
  success: boolean;
  message: string;
}

export type GetUserCardsParams = {
  page?: number;
  limit?: number;
};

export type GetAdminCardsParams = {
  search?: string;
  page?: number;
  limit?: number;
};