# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: Visiting Card Information Extractor API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication operations
  - name: extract
    description: Card extraction operations
  - name: cards
    description: User card operations
  - name: admin
    description: Admin operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/signup:
    post:
      operationId: authSignup
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/login:
    post:
      operationId: authLogin
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/me:
    get:
      operationId: authMe
      tags: [auth]
      summary: Get current logged-in user info
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /extract:
    post:
      operationId: extractCard
      tags: [extract]
      summary: Extract information from a visiting card image
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExtractCardRequest"
      responses:
        "200":
          description: Extraction successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExtractCardResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /cards:
    get:
      operationId: getUserCards
      tags: [cards]
      summary: Get current user's own extracted cards
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: number
        - in: query
          name: limit
          schema:
            type: number
      responses:
        "200":
          description: Cards retrieved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminCardsResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /cards/{id}:
    delete:
      operationId: deleteUserCard
      tags: [cards]
      summary: Delete one of the current user's own cards
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/cards:
    get:
      operationId: getAdminCards
      tags: [admin]
      summary: Get all extracted cards (admin only)
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: number
        - in: query
          name: limit
          schema:
            type: number
      responses:
        "200":
          description: Cards retrieved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminCardsResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/cards/{id}:
    delete:
      operationId: deleteAdminCard
      tags: [admin]
      summary: Delete any card (admin only)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Card deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Card not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/users:
    get:
      operationId: getAdminUsers
      tags: [admin]
      summary: Get all users (admin only)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Users retrieved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUsersResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    SignupRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
      required:
        - name
        - email
        - password

    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password

    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/UserProfile"
      required:
        - token
        - user

    UserProfile:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [user, admin]
        createdAt:
          type: string
      required:
        - id
        - name
        - email
        - role
        - createdAt

    CardData:
      type: object
      properties:
        name:
          type: string
        phones:
          type: array
          items:
            type: string
        emails:
          type: array
          items:
            type: string
        company:
          type: string
        designation:
          type: string
        address:
          type: string
        website:
          type: string
      required:
        - name
        - phones
        - emails
        - company
        - designation
        - address
        - website

    ExtractCardRequest:
      type: object
      properties:
        frontImage:
          type: string
        backImage:
          type: string
      required:
        - frontImage

    ExtractCardResponse:
      type: object
      properties:
        id:
          type: string
        data:
          $ref: "#/components/schemas/CardData"
        createdAt:
          type: string
      required:
        - id
        - data
        - createdAt

    CardRecord:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        userName:
          type: string
        data:
          $ref: "#/components/schemas/CardData"
        frontImageUrl:
          type: string
        backImageUrl:
          type: string
        createdAt:
          type: string
      required:
        - id
        - data
        - createdAt

    AdminCardsResponse:
      type: object
      properties:
        cards:
          type: array
          items:
            $ref: "#/components/schemas/CardRecord"
        total:
          type: number
        page:
          type: number
        totalPages:
          type: number
      required:
        - cards
        - total
        - page
        - totalPages

    AdminUserRecord:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          type: string
        cardCount:
          type: number
        createdAt:
          type: string
      required:
        - id
        - name
        - email
        - role
        - cardCount
        - createdAt

    AdminUsersResponse:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/AdminUserRecord"
        total:
          type: number
      required:
        - users
        - total

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message

    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message