# lib/api-zod/src/generated/types/adminUsersResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Visiting Card Information Extractor API
 * OpenAPI spec version: 0.2.0
 */
import type { AdminUserRecord } from "./adminUserRecord";

export interface AdminUsersResponse {
  users: AdminUserRecord[];
  total: number;
}