# lib/api-zod/src/generated/types/extractCardResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Visiting Card Information Extractor API
 * OpenAPI spec version: 0.2.0
 */
import type { CardData } from "./cardData";

export interface ExtractCardResponse {
  id: string;
  data: CardData;
  createdAt: string;
}